Updated and fixed tests
