+++ composer.json
@@ -32,8 +32,8 @@
     "csfix": ["php-cs-fixer fix"]
   },
   "require-dev": {
-    "pestphp/pest": "^1.0",
-    "minicli/pest-plugin-curly": "^0.1.1"
+    "pestphp/pest": "^2.0",
+    "minicli/pest-plugin-curly": "^0.2.1"
   },
   "config": {
     "allow-plugins": {
