+++ tests/Feature/Command/HelpCommandTest.php
@@ -10,7 +10,3 @@ test('default command "help" is correctly loaded', function () {
     $app->runCommand(['librarian', 'help']);
 })->expectOutputRegex("/help/");
 
-test('the "help test" command echoes command parameters', function () {
-    $app = getApp();
-    $app->runCommand(['librarian', 'help', 'test', 'user=erika']);
-})->expectOutputRegex("/erika/");
